+++ finetuner/__init__.py
@@ -48,15 +48,6 @@ def login(force: bool = False, interactive: Optional[bool] = None):
     ft.login(force=force, interactive=interactive)
 
 
-def notebook_login(force: bool = False):
-    warnings.warn(
-        message='Function `notebook_login` will be deprecated from Finetuner 0.7.0,'
-        'please use `login(interactive=True)` instead.',
-        category=DeprecationWarning,
-    )
-    ft.login(force=force, interactive=True)
-
-
 def list_callbacks() -> Dict[str, callback.CallbackStubType]:
     """List available callbacks."""
     return {