+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Remove `cpu` parameter from `create_run` function. ([#631](https://github.com/jina-ai/finetuner/pull/631))
 
+- Remove `notebook_login` function. ([#631](https://github.com/jina-ai/finetuner/pull/631))
+
 ### Changed
 
 ### Fixed